Fixed issue with .NET SDK package clashes when building or installing sample game (#742)
* Update Dockerfile to resolve .NET package clash

dotnet-sdk packages are contained in both amazonlinux and microsoft-prod repos causing issues when building Sample Game. Added package exclusions to amazonlinux.repo to resolve.

* Update install.sh to resolve .NET package clash

dotnet-sdk packages are present in both amazonlinux and microsoft-prod repos which causes a package clash when building or installing the sample game. Added package exclusions to the amazonlinux repo to resolve.